refactor: GH-37 Moving site to its own directory
